renamed to http_client and http_server
