+++ tests/rest_test/rest_server/rest_server.cpp
@@ -7,7 +7,7 @@
 #include <sstream>
 
 #include "glaze/glaze.hpp"
-#include "glaze/rest/rest_server.hpp"
+#include "glaze/rest/http_server.hpp"
 #include "glaze/rpc/registry.hpp"
 
 struct User
